Fixed bug in `QuickProgram.reset_phase()` when default registers are used.
When default registers were used, the parameters for playing dummy pulses that trigger the phase reset were not set properly, which leads to a non-zero pulse played at `reset_phase()`. This was because I forgot to also temporarily empty the `AbsRegisterManager.default_regs` before setting those dummy parameters. This is fixed now.